Pull out PostgresConnectionManager
